🚚 Move contextIcons
