+++ static/app/components/events/contexts/contextIcon.tsx
